feat(scheduler): 渠道定时检测间隔支持小数，允许分钟级精度
将 channel_check_interval_hours 类型从 int 改为 float，0.5 即 30 分钟。
DB 迁移自动将旧 int 行升级为 float；前端 Number() 原本即可处理小数，无需额外改动。

+++ internal/app/server.go
@@ -250,12 +250,12 @@ func NewServer(store storage.Store) *Server {
 	s.startBackgroundWorkers()
 
 	channelCheckIntervalHours := normalizeChannelCheckIntervalHours(
-		configService.GetInt("channel_check_interval_hours", defaultChannelCheckIntervalHours),
+		configService.GetFloat("channel_check_interval_hours", defaultChannelCheckIntervalHours),
 	)
 	if channelCheckIntervalHours == 0 {
 		log.Print("[INFO] 渠道定时检测未启用（channel_check_interval_hours=0）")
 	} else {
-		s.startScheduledChannelCheckLoop(time.Duration(channelCheckIntervalHours) * time.Hour)
+		s.startScheduledChannelCheckLoop(time.Duration(channelCheckIntervalHours * float64(time.Hour)))
 	}
 
 	return s
